bug fixed and hanling error

+++ services/userService.js
@@ -20,16 +20,11 @@ export const getAllUsers = async () => {
 };
 
 export const update = async (user) => {
-  console.log(user)
   const id = user._id;
   const User = models.User;
   let model = await User.findById(id);
-  console.log('model',model)
 
-  // model.username = user.username;
-  // model.save();
-
-  if(model) {
+  if (model) {
     model.username = user.username;
     model.save();
     return model;
@@ -38,18 +33,15 @@ export const update = async (user) => {
   return null;
 };
 
-
-
 export const deleteById = async (id) => {
   //remove->obsolote
   //delete one
   //delte many
   const User = models.User;
-  const result = await User.deleteOne({_id:id});
-  return result;
-  /**
-   *  if the user is not null,delete it
-   *  return user not found
-   *  
-   */ 
+  let model = await User.findById(id);
+  if (model){
+    let result = await User.deleteOne({_id:id});
+    return result;
+  }
+  return new Error("User not found by the id: " + id) 
 };
